+++ game-frontend/src/app/service/planet.service.ts
@@ -22,11 +22,20 @@ export class PlanetService extends GameBaseService {
     this.findMyPlanets();
   }
 
-  public findMyPlanets(): void {
-    this.doGetWithAuthorizationToGame('planet/findMyPlanets').subscribe(result => this._myPlanets.next(result));
+  public findMyPlanets(): Promise<void> {
+    return new Promise(resolve => {
+      this.doGetWithAuthorizationToGame('planet/findMyPlanets').subscribe(result => {
+        this._myPlanets.next(result);
+        resolve();
+      });
+    });
   }
 
   public isMine(planet: PlanetPojo): boolean {
     return planet.ownerId === this._loginSessionService.findTokenData().id;
+  }
+
+  public leavePlanet(planet: PlanetPojo): Observable<void> {
+    return this._doPostWithAuthorizationToGame('planet/leave?planetId=' + planet.id, {});
   }
 }